chore: useMediaQuery to determine if view is in mobile
If in mobile then setOpen state to false, add a hamburger menu icon that
is visible on mobile screens to show/hide the menu. Also add an overlay
to the sidebar that shows when in mobile screen and sidebar menu is
open. Add click event to overlay to close the sidebar menu.